Spot Light Made and Light Flicker script
Set up a spot light connected to the character's camera to allow it to move along while looking around. Also made a script for the light flickering, though at the moment it might be nonfunctional (it seems to take a long time to load, and my task manager keeps marking as "not responding" when I check on it to see if something's up) and as such isn't attached to the light at the moment.